Changed radius to int

+++ include/entity.h
@@ -16,7 +16,7 @@ typedef struct Entity_S
     float       frame;              /**<Current frame of sripte*/
     float       maxFrames;          /**<Maximum number of frames in a sprite*/
     Vector2D    position;           /**<2D position of entity*/
-    float       radius;             /**<how wide this entity is*/
+    int         radius;             /**<how wide this entity is*/
     Vector2D    drawOffset;         /**<Offset of collider*/
     Vector2D    size;               /**<Size vector*/
